requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
supabase==2.7.4
openai==1.45.0
anthropic==0.34.0
apscheduler==3.10.4
httpx==0.27.2
python-telegram-bot==21.5
duckduckgo-search==6.3.0
playwright==1.47.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
python-dotenv==1.0.1
pydantic==2.9.2
google-auth
google-auth-oauthlib
google-api-python-client
google-auth-httplib2